refactor: same Server Component wrapper for the remaining routes
Apply the same split to /contact, /playx1, and /banca-do-ingresso so
their page.tsx exports are Server Components that render the existing
client UI (ContactContent, PlayX1Content, BancaDoIngressoContent).

The interactive parts (form state, copy-to-clipboard, image modal,
window.open) stay client, as the issue requires. Build and route
bundle sizes are unchanged from the previous commit.
